+++ R/omop_relations.R
@@ -1,4 +1,6 @@
-#' find omop concept relations of one passed - immediate relations with indication of relationship
+#' recursively extract omop concept relations of one passed - immediate relations with indication of relationship
+#'
+#' TODO need to sort repeated rows
 #'
 #' @param c_id single omop concept_id or exact concept_name to get relations of, default NULL returns all
 #' @param c_ids one or more concept_id to filter by, default NULL for all
@@ -9,85 +11,101 @@
 #' @param r_ids one or more relationship_id to filter by, default NULL for all, e.g c('Is a','Subsumes')
 #' @param itself whether to include relations to concept itself, default=FALSE
 #' @param names2avoid concept names to avoid, defaults to generic concepts with lots relations, can be set to NULL
-#' @param join_names whether to join concept names onto ids, default TRUE, FALSE used by recursive to speed up
 #' @param messages whether to print info messages, default=TRUE
+#' @param nsteps number of recursions to search
+#' @param add_step_column whether to add column with step level, default=TRUE
 #' @return a dataframe of concepts and attributes
 #' @export
 #' @examples
 #' omop_relations("Non-invasive blood pressure")
-#' omop_relations("Non-invasive blood pressure",r_ids=c('Is a','Subsumes'))
+#' omop_relations("Non-invasive blood pressure",nsteps=2,r_ids=c('Is a','Subsumes'))
 #' #omop_relations("lenalidomide")
 #' #omop_relations(1633308)
 #' #chemodrugs <- omop_relations("Cytotoxic chemotherapeutic",v_ids="HemOnc",d_ids="Regimen")
 #' #cmde <- omop_relations(v_ids="Cancer Modifier")
 omop_relations <- function(c_id=NULL,
-                                c_ids=NULL,
-                                d_ids=NULL,
-                                v_ids=NULL,
-                                cc_ids=NULL,
-                                standard=NULL,
-                                r_ids=NULL,
-                                itself=FALSE,
-                                names2avoid=c("SNOMED CT core","Defined","Primitive"),
-                                join_names = TRUE,
-                                messages=TRUE) {
+                           c_ids=NULL,
+                           d_ids=NULL,
+                           v_ids=NULL,
+                           cc_ids=NULL,
+                           standard=NULL,
+                           r_ids=NULL,
+                           itself=FALSE,
+                           names2avoid=c("SNOMED CT core","Defined","Primitive"),
+                           messages=TRUE,
+                           nsteps=1,
+                           add_step_column=TRUE) {
 
 
   #checks c_id and gets name (ALL if c_id==NULL)
   res <- check_c_id(c_id)
-  c_id <- res$c_id[1]
+  c_id0 <- res$c_id[1]
   name1 <- res$name1[1]
 
-  # default avoidance of generic concepts that have huge num relations
-  if (name1 %in% names2avoid)
+  if (messages) message("recursively querying concept relations of: ",name1," - may take more than a few seconds")
+
+  dfall <- NULL
+
+  for(step in 1:nsteps)
   {
-    if (messages) message("not returning relations of ",name1," because lots of, you can change by setting names2avoid to NULL or subset of default")
-    return(NULL)
+
+    if (messages) message("step ",step," of ",nsteps)
+
+    # get relations of each concept from the previous level
+    if (step == 1) {
+      prev_c_ids <- c_id0
+    } else {
+      prev_c_ids <- unique(dfprev$concept_id_2)
+    }
+
+    dfprev <- NULL
+
+    for(c_id in prev_c_ids) {
+
+      # get immediate relations
+      dfprev1 <- omop_relations1step(c_id=c_id,
+                               c_ids=c_ids,
+                               d_ids=d_ids,
+                               v_ids=v_ids,
+                               cc_ids=cc_ids,
+                               standard=standard,
+                               r_ids=r_ids,
+                               itself=itself,
+                               names2avoid=names2avoid,
+                               messages=messages,
+                               join_names=FALSE)
+
+      #add step column that can be used in plot colouring
+      if (add_step_column)
+      {
+        dfprev1 <- dfprev1 |> mutate(step=step)
+        #need a relation to itself to get colouring right
+        #in some options standard one gets filtered elsewhere
+        #'Is' is a non standard relationship_id
+        if (step == 1)
+        {
+          r_to_itself <- dfprev1 |>
+            head(1) |>
+            mutate( concept_id_1=c_id0,
+                    concept_id_2=c_id0,
+                    relationship_id="Is",
+                    step=0)
+
+          dfprev1 <- bind_rows(r_to_itself, dfprev1)
+        }
+      }
+      dfprev <- bind_rows(dfprev,dfprev1)
+    }
+
+    dfall <- bind_rows(dfall,dfprev)
+
   }
 
-  #spaces before make it clearer when running recursively
-  if (messages) message("   querying concept relations of: ",name1," - may take a few seconds")
+  #join names onto ids only at end to make faster
+  dfall <- dfall |> omop_join_name_all()
 
-  df1 <- omopcept::omop_concept_relationship()
-
-  if (c_id != "none") df1 <- df1 |>
-    filter(concept_id_1 == c_id) #TODO check whether I want to add concept_id_2 to this filter
-    #TODO add %in% c_id to vectorise
-
-  if (!is.null(r_ids)) {
-    df1 <- df1 |>
-      filter(relationship_id %in% r_ids)
-  }
-
-  #get most attributes from omop_concept() to join
-  #not all e.g.valid_start_date, end & invalid_reason already in relations
-  concept_attributes <- omopcept::omop_concept() |>
-    select(!c(concept_name,valid_start_date, valid_end_date, invalid_reason))
-
-  df1 <- df1 |>
-    left_join(concept_attributes, by = dplyr::join_by(concept_id_2 == concept_id)) |>
-    omop_filter_concepts(c_ids=c_ids, d_ids=d_ids, v_ids=v_ids, cc_ids=cc_ids, standard=standard)
-
-  #joining names to concept ids, could make it optional & not do in recursive
-  if (join_names)
-  {
-    df1 <- df1 |>  omop_join_name_all()
-  }
-
-  # default option remove relations to itself
-  if (itself==FALSE) df1 <- df1 |> filter(concept_id_1 != concept_id_2)
-
-  #old move name column next to id to make more readable
-  #not needed now with way names are joined
-  #dplyr::relocate(concept_name_2, .after=concept_id_1) |>
-  df1 <- df1 |> collect()
-
-  if (messages) message("returning ",nrow(df1)," concepts")
-
-  return(df1)
-
+  return(dfall)
 }
-
 
 #' super short name func to find relations
 #' @rdname omop_relations